Move "text" utils to own file.
