+++ src/client/refactor/proxy.ts
@@ -7,7 +7,8 @@ import '../common/extensions';
 import { createDeferred, Deferred } from '../common/helpers';
 import { IPythonExecutionFactory } from '../common/process/types';
 import { IPythonSettings } from '../common/types';
-import { getWindowsLineEndingCount, IS_WINDOWS } from '../common/util';
+import { IS_WINDOWS } from '../common/util';
+import { getWindowsLineEndingCount } from '../common/utils/text';
 import { IServiceContainer } from '../ioc/types';
 
 export class RefactorProxy extends Disposable {
